requirements.txt
python
django
django-crispy-forms
django_filter
djangorestframework
postgresql
postgresql-contrib
